% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.msea.R
\name{barplot}
\alias{barplot}
\title{plot count bar on each meatbolite-set}
\usage{
barplot(x, col = "cm.colors", show.limit = 20)
}
\arguments{
\item{x}{a msea object}

\item{col}{color scheme}

\item{show.limit}{the number of metabolite-sets to plot}
}
\value{
plot
}
\description{
plot count bar on each meatbolite-set
}
\examples{
library(MSEApdata)
data(mset_SMPDB_Metabolic_format_HMDB)
data(msea.example)
res <- msea(mset_SMPDB_Metabolic_format_HMDB, msea.example)
barplot(res)
}
