% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.msea.R
\name{dotplot}
\alias{dotplot}
\title{plot count-sized dot on (x, y) = (metabolite-sets, pvalue)}
\usage{
dotplot(x, show.limit = 20)
}
\arguments{
\item{x}{A msea object}

\item{show.limit}{the number of metabolite-sets to plot}
}
\value{
plot
}
\description{
plot count-sized dot on (x, y) = (metabolite-sets, pvalue)
}
\examples{
library(MSEApdata)
data(mset_SMPDB_Metabolic_format_HMDB)
data(msea.example)
res <- msea(mset_SMPDB_Metabolic_format_HMDB, msea.example)
dotplot(res)
}
