% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fukushima17_INC}
\alias{fukushima17_INC}
\title{list of signicant metabolite set based on Fukushima et al. (2017).}
\description{
A metabolite list. Using GC-TOF/MS and LC-q-TOF/MS, We analyzed the
Arabidopsis metabolome of the aerial parts of individual WT plants
under mild oxidative stress elicited by methyl viologen.
}
\details{
This list is corresponding to significantly accumulated metabolites
between mild oxidative stress and control conditions.
This study identified 37 significant increased metabolites
(log2foldchange >= 1 and FDR < 0.05) including
primary- and secondary metabolites, mainly flavonols and anthocyanins.
}
\references{
Atsushi Fukushima, Mami Iwasa et al. Front Plant Sci 8:1464 (2017) doi: 10.3389/fpls.2017.01464
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
