% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgml.import.R
\name{kgml.import}
\alias{kgml.import}
\title{Extraction of compound set from KGML file (*.xml) to list}
\usage{
kgml.import(filename)
}
\arguments{
\item{filename}{a file name of KEGG pathway in KGML format}
}
\value{
list of KEGG compound IDs within a pathway
}
\description{
Convertion of KEGG's XML (KGML) file (e.g. ath00630.xml) to compound set list
}
\details{
The KGML files should be downloaded from KEGG FTP according to your own
licence.
}
\examples{

    filename <- system.file("extdata/hsa04010.xml",
                            package="KEGGgraph") ## KGML downloaded from KEGG ftp
    kgml.import(filename)

}
\author{
Atsushi Fukushima
}
\keyword{list}
