% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kusano}
\alias{kusano}
\title{list of significant metabolites based on Kusano et al. Plant J (2011)}
\description{
A metabolite list. We analyzed the Arabidopsis metabolome of the aerial
regions of individual WT plants under UV-B treatment
by GC-TOF/MS.
}
\details{
This list is corresponding to significantly accumulated metabolites
between UV-B (1 day treatment) and control conditions.
}
\references{
Miyako Kusano et al. Plant J 67:354-369 (2011)
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
