% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msea.example}
\alias{msea.example}
\title{An example INPUT list based on HMDB-IDs}
\description{
An example INPUT metabolite-set related to MSEA analysis in MetaboAnalyst.
}
\references{
Xia et al. Nucleic Acids Res. 43:W251-W257 (2015).
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
