% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteset.enrichment.R
\name{msea}
\alias{msea}
\title{identifies biologically meaningful accumulations in metabolomic data}
\usage{
msea(metabolite.set, mets.queryset, assessment = "E", midp = FALSE)
}
\arguments{
\item{metabolite.set}{a list of metabolite sets}

\item{mets.queryset}{a list of metabolites for which enrichment test is to be done}

\item{assessment}{a kind of alternative test (E, enrichment; D, depletion; and ED, enrichment or depletion)}

\item{midp}{logical. TRUE gives mid p-values}
}
\value{
msea object
}
\description{
This function performs Metabolite Set
Enrichment Analysis (MSEA) with KEGG COMPOUND IDs.
}
\examples{
data(mset_SMPDB_format_KEGG)
data(kusano)

res <- msea(mset_SMPDB_format_KEGG, kusano)

}
\author{
Atsushi Fukushima
}
