% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mset10hmdb}
\alias{mset10hmdb}
\title{An example UNIVERSE list based on HMDB-IDs}
\description{
A selected metabolite-set in MetaboAnalyst.
}
\details{
It includes 10 metabolite-sets selected from the pre-defined metabolite
sets in MetaboAnalyst (filename: metaboliteset_example.csv).
}
\references{
Xia et al. Nucleic Acids Res. 43:W251-W257 (2015).
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
