% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mset_AraCyc_format_HMDB}
\alias{mset_AraCyc_format_HMDB}
\title{A metabolite set (HMDB-IDs) based on AraCyc}
\description{
A metabolite set based on AraCyc
(http://www.plantcyc.org/databases/aracyc/15.0).
We extracted the pathway information from the biopax-level2.owl
AraCyc pathway IDs and HMDB IDs were extracted. This dataset
contains 350 metabolite sets linked to HMDB IDs. There are only
metabolite sets containing at least 5 compounds.
}
\references{
Zhang P, et al. Plant Physiol. 138:27-37 (2005).
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
