% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mset_AtMetExpress_Flavonoids_format_KNApSAcK}
\alias{mset_AtMetExpress_Flavonoids_format_KNApSAcK}
\title{A metabolite set (KNApSAcK-IDs) based on AtMet}
\description{
A metabolite set based on AtMetExpress, has been curated manually.
We extracted the metabolite set information from publicly available
metabolite datasets involved in flavonoid pathways in Arabidopsis.
KNApSAcK IDs were assigned to the sets. This dataset contains
4 metabolite sets linked to KNApSAcK IDs. There are only
metabolite sets containing at least 5 compounds.
}
\references{
Saito et al. Plant Physiol Biochem. 72:21-34 (2013).
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
