% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mset_AtMetExpress_Stress_format_KEGG}
\alias{mset_AtMetExpress_Stress_format_KEGG}
\title{A metabolite set (KEGG COMPOUND-IDs) based on AtMet}
\description{
A metabolite set based on AtMetExpress, has been curated manually.
We extracted the metabolite set information from publicly available
metabolite datasets in response to various stress treatment.
KEGG IDs were assigned to the sets. This dataset contains
9 metabolite sets linked to KEGG COMPOUND IDs. There are only
metabolite sets containing at least 5 compounds.
}
\references{
Obata and Fernie, Cell Mol Life Sci 69:3225-3243 (2012).
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
