% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mset_EcoCyc_format_KEGG}
\alias{mset_EcoCyc_format_KEGG}
\title{A metabolite set (KEGG COMPOUND-IDs) based on EcoCyc}
\description{
A metabolite set based on EcoCyc version 21.5
(https://ecocyc.org/).
We extracted the pathway information from the biopax-level2.owl
EcoCyc pathway IDs and KEGG COMPOUND IDs were extracted. This dataset
contains 222 metabolite sets linked to KEGG COMPOUND IDs. There are only
metabolite sets containing at least 5 compounds.
}
\references{
Keseler IM, et al. Nucleic Acids Res. 2017 Jan 4;45(D1):D543-D550.
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
