% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mset_FlyCyc_format_KEGG}
\alias{mset_FlyCyc_format_KEGG}
\title{A metabolite set (KEGG COMPOUND-IDs) based on FlyCyc}
\description{
A metabolite set based on FlyCyc version 21.5
(https://biocyc.org/FLY/organism-summary?object=FLY).
We extracted the pathway information from the biopax-level2.owl
FlyCyc pathway IDs and KEGG COMPOUND IDs were extracted. This dataset
contains 90 metabolite sets linked to KEGG COMPOUND IDs. There are only
metabolite sets containing at least 5 compounds.
}
\references{
Kenneth L Wiley, Harvard University; Andy Schroeder, FlyBase/Harvard University; William Gelbart, FlyBase/Harvard University
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
