% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mset_HumanCyc_format_KEGG}
\alias{mset_HumanCyc_format_KEGG}
\title{A metabolite set (KEGG COMPOUND-IDs) based on HumanCyc}
\description{
A metabolite set based on HumanCyc version 21.5
(https://humancyc.org/).
We extracted the pathway information from the biopax-level2.owl
HumanCyc pathway IDs and KEGG COMPOUND IDs were extracted. This dataset
contains 89 metabolite sets linked to KEGG COMPOUND IDs. There are only
metabolite sets containing at least 5 compounds.
}
\references{
Trupp M, et al. Genome Biol 2010, 11(Suppl 1):O12
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
