% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mset_MetaCyc_format_HMDB}
\alias{mset_MetaCyc_format_HMDB}
\title{A metabolite set (HMDB IDs) based on MetaCyc}
\description{
A metabolite set based on MetaCyc version 21.5
(https://MetaCyc.org).
We extracted the pathway information from the biopax-level2.owl
MetaCyc pathway IDs and HMDB IDs were extracted. This dataset
contains 2059 metabolite sets linked to HMDB IDs. There are only
metabolite sets containing at least 5 compounds.
}
\references{
Caspi R et al. Nucleic Acids Res, 2018 Jan 4;46(D1):D633-D639.
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
