% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mset_MouseCyc_format_KEGG}
\alias{mset_MouseCyc_format_KEGG}
\title{A metabolite set (KEGG COMPOUND-IDs) based on MouseCyc}
\description{
A metabolite set based on MouseCyc version 1.7
(http://mousecyc.jax.org/MOUSE/organism-summary).
We extracted the pathway information from the biopax-level2.owl
MouseCyc pathway IDs and KEGG COMPOUND IDs were extracted. This dataset
contains 71 metabolite sets linked to KEGG COMPOUND IDs. There are only
metabolite sets containing at least 5 compounds.
}
\references{
Evsikov AV, et al. Genome Biol 2009, 10:R84
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
