% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mset_PlantCyc_format_HMDB}
\alias{mset_PlantCyc_format_HMDB}
\title{A metabolite set (HMDB-IDs) based on PlantCyc}
\description{
A metabolite set based on PlantCyc 12.0 (https://www.plantcyc.org/).
We extracted the pathway information from the biopax-level2.owl
PlantCyc pathway IDs and HMDB IDs were extracted. This dataset
contains 714 metabolite sets linked to HMDB IDs. There are only
metabolite sets containing at least 5 compounds.
}
\references{
Zhang P, et al. Plant Physiol. 153:1479-91 (2010).
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
