% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mset_SMPDB_format_KEGG}
\alias{mset_SMPDB_format_KEGG}
\title{A metabolite set (KEGG COMPOUND-IDs) based on SMPDB}
\description{
A metabolite set based on SMPDB (http://smpdb.ca/).
We downloaded the CSV file of metabolite names that
linked to SMPDB pathways (ver. 2.0). SMPDB pathway IDs and
KEGG COMPOUND IDs were extracted.
}
\references{
Jewison T, et al. Nucleic Acids Res. 42:D478-484 (2014).
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
