% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mset_SMPDB_Metabolic_format_HMDB}
\alias{mset_SMPDB_Metabolic_format_HMDB}
\title{A metabolite set (HMDB-IDs) based on SMPDB core metabolic pathways}
\description{
A metabolite set based on SMPDB (http://smpdb.ca/).
We downloaded the CSV file of metabolite names that
linked to SMPDB pathways (ver. 2.0). SMPDB pathway IDs and
HMDB IDs were extracted. This dataset includes 92 core metabolic pathways.
}
\references{
Jewison T, et al. Nucleic Acids Res. 42:D478-484 (2014).
}
\author{
Atsushi Fukushima
}
\keyword{datasets}
