% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.msea.R
\name{netplot}
\alias{netplot}
\title{plot msea result with network}
\usage{
netplot(x, edgetable, show.limit = 20)
}
\arguments{
\item{x}{A msea result}

\item{edgetable}{A csv generated with write.network function}

\item{show.limit}{The number of metabolite-sets to plot}
}
\description{
plot msea result with network
}
\examples{
data(kusano)
data(mset_SMPDB_format_KEGG)
res <- msea(mset_SMPDB_format_KEGG, kusano)
write.network(mset_SMPDB_format_KEGG, shared.metabolite = 20)
netplot(res, "./mset_SMPDB_format_KEGG_edges_share20.csv")
}
\keyword{internal}
