% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gmt.R
\name{read.gmt}
\alias{read.gmt}
\title{parse gmt file to a list format}
\usage{
read.gmt(file, format = "KEGG")
}
\arguments{
\item{file}{*.gmt file (a TAB-delimited file)}

\item{format}{select the compound DB-IDs ("KEGG", "HMDB", "KNApSacK")}
}
\value{
list
}
\description{
This function parses gmt file to a list format
for enrichment analysis.
}
\examples{

file <- system.file("extdata", "sample_SMPDB.gmt", package = "MSEAp")
smp <- read.gmt(file)

}
\author{
Atsushi Fukushima
}
