% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.wikipathways.R
\name{read.wikipathways}
\alias{read.wikipathways}
\title{parse Wikipathways biopax file to a list format}
\usage{
read.wikipathways(file)
}
\arguments{
\item{file}{*.biopax file}
}
\value{
list
}
\description{
This function parses wikipathways biopax file to a list format
for enrichment analysis.
}
\examples{

file <- system.file("extdata", "WP3620_95486.owl", package = "MSEAp")
wikip <- read.wikipathways(file)

}
\author{
Kozo Nishida
}
