% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.KEGG.gmt.R
\name{setup.KEGG.gmt}
\alias{setup.KEGG.gmt}
\title{Setup metabolite set by KEGG pathways}
\usage{
setup.KEGG.gmt(organism = "ath", listformat = TRUE, target.path = NULL,
  debug = TRUE)
}
\arguments{
\item{organism}{3-letters KEGG org code (e.g., eco, hsa, and ath)}

\item{listformat}{if true, return list object}

\item{target.path}{a character vector of KEGG pahtway IDs (default = NULL)}

\item{debug}{if true, debug mode run}
}
\value{
list or data frame objects
}
\description{
This function generates metabolite set for metabolite
enrichment analysis using the KEGG REST API.
}
\examples{
## debug mode
res <- setup.KEGG.gmt(listformat = FALSE)
head(res)

}
\author{
Atsushi Fukushima
}
