% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupPathwayClass.R
\name{setupPathwayClass}
\alias{setupPathwayClass}
\title{Convert metabolite set (multiple) / tar.gz to list}
\usage{
setupPathwayClass(filepath)
}
\arguments{
\item{filepath}{file path of KEGG's tar.gz files}
}
\value{
list of KEGG compound IDs within a pathway
}
\description{
Convertion of KEGG's tar.gz files (e.g. hsa.tar.gz) to metabolite set list
}
\details{
The tar.gz files should be downloaded from KEGG FTP according to your own
licence.
}
\examples{

	## filepath <- "./hsa.tar.gz" ## *.tar.gz downloaded from KEGG
	## res <- setupPathwayClass(filepath) ## convert to list

}
\author{
Atsushi Fukushima
}
\keyword{list}
