% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{supported.msets}
\alias{supported.msets}
\title{Display supported metabolite-sets for metabolite set enrichment analysis (MSEA).}
\usage{
supported.msets()
}
\value{
metabolite-set name available in MSEAp package.
}
\description{
Display supported metabolite-sets for metabolite set enrichment analysis (MSEA).
}
\examples{
supported.msets()
}
