% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.msea.R
\name{netplot}
\alias{netplot}
\title{plot msea result with network}
\usage{
netplot(x, mset, shared.metabolite = 3, show.limit = 20,
  sendto = c("visnetwork", "cytoscape"))
}
\arguments{
\item{x}{A msea result}

\item{mset}{A list of metabolite-sets}

\item{shared.metabolite}{The number of shared metabolites to connect the metabolite-set nodes with edges}

\item{show.limit}{The number of metabolite-sets to plot}

\item{sendto}{The target of the network visualization}
}
\description{
plot msea result with network
}
\examples{
library(MSEApdata)
data(kusano)
data(mset_SMPDB_format_KEGG)
res <- msea(mset_SMPDB_format_KEGG, kusano)
netplot(res, mset_SMPDB_format_KEGG, shared.metabolite = 20)
# You can also send the network to Cytoscape with RCy3
# netplot(res, mset_SMPDB_format_KEGG, shared.metabolite = 20, sendto = "cy")
}
\keyword{internal}
