% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.msea.R
\name{write.network}
\alias{write.network}
\title{write network csv files for a metabolite-set}
\usage{
write.network(mset, shared.metabolite = 3)
}
\arguments{
\item{mset}{A list of metabolite-sets}

\item{shared.metabolite}{The number of shared metabolites to connect the metabolite-set nodes with edges}
}
\description{
write network csv files for a metabolite-set
}
\examples{
data(mset_SMPDB_format_KEGG)
write.network(mset_SMPDB_format_KEGG, shared.metabolite = 20)
}
